Move CacheInvalidationOptimizer to separate file and cleanup row_id_is_last_column (#27)
