+++ CMakeLists.txt
@@ -16,6 +16,7 @@ endif()
 include_directories(src/include)
 
 set(EXTENSION_SOURCES
+    src/cache_invalidation_optimizer.cpp
     src/logical_cache_invalidator.cpp
     src/physical_cache_invalidator.cpp
     src/query_condition_cache_extension.cpp
